fix: Address feedback
