+++ src/auth/auth-api-request.ts
@@ -41,8 +41,8 @@ import {
   OIDCConfigServerRequest, SAMLConfigServerRequest, AuthProviderConfig,
   OIDCUpdateAuthProviderRequest, SAMLUpdateAuthProviderRequest,
 } from './auth-config';
-import { Tenant, TenantOptions } from './tenant';
-import { TenantServerResponse } from './tenant-internal';
+import { TenantOptions } from './tenant';
+import { TenantServerResponse, TenantUtils } from './tenant-internal';
 
 
 /** Firebase Auth request header. */
@@ -1808,7 +1808,7 @@ const UPDATE_TENANT = new ApiSettings('/tenants/{tenantId}?updateMask={updateMas
   .setResponseValidator((response: any) => {
     // Response should always contain at least the tenant name.
     if (!validator.isNonEmptyString(response.name) ||
-          !Tenant.getTenantIdFromResourceName(response.name)) {
+          !TenantUtils.getTenantIdFromResourceName(response.name)) {
       throw new FirebaseAuthError(
         AuthClientErrorCode.INTERNAL_ERROR,
         'INTERNAL ASSERT FAILED: Unable to update tenant',
@@ -1843,7 +1843,7 @@ const CREATE_TENANT = new ApiSettings('/tenants', 'POST')
   .setResponseValidator((response: any) => {
     // Response should always contain at least the tenant name.
     if (!validator.isNonEmptyString(response.name) ||
-          !Tenant.getTenantIdFromResourceName(response.name)) {
+          !TenantUtils.getTenantIdFromResourceName(response.name)) {
       throw new FirebaseAuthError(
         AuthClientErrorCode.INTERNAL_ERROR,
         'INTERNAL ASSERT FAILED: Unable to create new tenant',
@@ -1961,7 +1961,7 @@ export class AuthRequestHandler extends AbstractAuthRequestHandler {
   public createTenant(tenantOptions: TenantOptions): Promise<TenantServerResponse> {
     try {
       // Construct backend request.
-      const request = Tenant.buildServerRequest(tenantOptions, true);
+      const request = TenantUtils.buildServerRequest(tenantOptions, true);
       return this.invokeRequestHandler(this.tenantMgmtResourceBuilder, CREATE_TENANT, request)
         .then((response: any) => {
           return response as TenantServerResponse;
@@ -1984,7 +1984,7 @@ export class AuthRequestHandler extends AbstractAuthRequestHandler {
     }
     try {
       // Construct backend request.
-      const request = Tenant.buildServerRequest(tenantOptions, false);
+      const request = TenantUtils.buildServerRequest(tenantOptions, false);
       // Do not traverse deep into testPhoneNumbers. The entire content should be replaced
       // and not just specific phone numbers.
       const updateMask = utils.generateUpdateMask(request, ['testPhoneNumbers']);
